added new suite
