+++ sql/catalyst/src/main/java/org/apache/spark/sql/connector/read/SupportsPushDownJoin.java
@@ -47,8 +47,8 @@ public interface SupportsPushDownJoin extends ScanBuilder {
    * @param other {@code SupportsPushDownJoin} that this {@code SupportsPushDownJoin}
    * gets joined with.
    * @param joinType the type of join.
-   * @param leftSideRequiredColumnWithAliases required output of the left side {@code SupportsPushDownJoin}
-   * @param rightSideRequiredColumnWithAliases required output of the right side {@code SupportsPushDownJoin}
+   * @param leftSideRequiredColumnsWithAliases required output of the left side {@code SupportsPushDownJoin}
+   * @param rightSideRequiredColumnsWithAliases required output of the right side {@code SupportsPushDownJoin}
    * @param condition join condition. Columns are named after the specified aliases in
    * {@code leftSideRequiredColumnWithAliases} and {@code rightSideRequiredColumnWithAliases}
    * @return True if join has been successfully pushed down.
@@ -56,8 +56,8 @@ public interface SupportsPushDownJoin extends ScanBuilder {
   boolean pushDownJoin(
       SupportsPushDownJoin other,
       JoinType joinType,
-      ColumnWithAlias[] leftSideRequiredColumnWithAliases,
-      ColumnWithAlias[] rightSideRequiredColumnWithAliases,
+      ColumnWithAlias[] leftSideRequiredColumnsWithAliases,
+      ColumnWithAlias[] rightSideRequiredColumnsWithAliases,
       Predicate condition
   );
